refactor: reorganize algebra module and improve error handling
- Split the algebra/element.rs file into separate point.rs and scalar.rs files
- Replace implicit Sum trait with explicit sum methods with error handling
- Update import paths across the codebase
- Add new Algebra error type to better handle errors during operations
