+++ src/crypto/primitives/algebra/scalar.rs
@@ -1,30 +1,12 @@
-use std::iter::Sum;
-
 use curv::{
     elliptic::curves::secp256_k1::Secp256k1 as CurvSecp256k1, elliptic::curves::Point as CurvPoint,
     elliptic::curves::Scalar as CurvScalar,
 };
 use serde::{Deserialize, Serialize};
 
-use crate::crypto::primitives::algebra::Scheme;
+use crate::crypto::primitives::algebra::{point::Point, Error, Scheme};
 
 type Result<T> = std::result::Result<T, Error>;
-
-#[derive(Debug)]
-#[derive(thiserror::Error)]
-pub enum Error {
-    #[error("{0}")]
-    Encode(#[from] bincode::error::EncodeError),
-
-    #[error("{0}")]
-    Decode(#[from] bincode::error::DecodeError),
-
-    #[error("Inconsistent variants")]
-    InconsistentVariants,
-
-    #[error("Iterator empty")]
-    IteratorEmpty,
-}
 
 #[derive(Clone)]
 #[derive(Debug)]
@@ -32,14 +14,6 @@ pub enum Error {
 #[derive(PartialEq, Eq)]
 pub enum Scalar {
     Secp256k1(CurvScalar<CurvSecp256k1>),
-}
-
-#[derive(Clone)]
-#[derive(Debug)]
-#[derive(Serialize, Deserialize)]
-#[derive(PartialEq, Eq)]
-pub enum Point {
-    Secp256k1(CurvPoint<CurvSecp256k1>),
 }
 
 impl Scalar {
@@ -127,72 +101,21 @@ impl Scalar {
             .map(|(s, _)| s)
             .map_err(Error::from)
     }
-}
 
-impl Point {
-    pub fn zero(scheme: Scheme) -> Self {
-        match scheme {
-            Scheme::Secp256k1 => Point::secp256k1_zero(),
-        }
-    }
+    pub fn sum<I: Iterator<Item = Self>>(iter: I) -> Result<Self> {
+        let mut iter = iter.peekable();
+        let mut sum = iter.next().ok_or(Error::IteratorEmpty)?;
 
-    pub fn secp256k1_zero() -> Self {
-        Point::Secp256k1(curv::elliptic::curves::Point::zero())
-    }
-
-    pub fn to_curv_secp256k1_point(&self) -> Result<CurvPoint<CurvSecp256k1>> {
-        match self {
-            Point::Secp256k1(point) => Ok(point.clone()),
-        }
-    }
-
-    pub fn is_same_type(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Point::Secp256k1(_), Point::Secp256k1(_)) => true,
-        }
-    }
-
-    pub fn is_same_type_scalar(&self, other: &Scalar) -> bool {
-        match (self, other) {
-            (Point::Secp256k1(_), Scalar::Secp256k1(_)) => true,
-        }
-    }
-
-    pub fn to_vec(&self) -> Result<Vec<u8>> {
-        bincode::serde::encode_to_vec(self, bincode::config::standard()).map_err(Error::from)
-    }
-
-    pub fn from_slice(bytes: &[u8]) -> Result<Self> {
-        bincode::serde::decode_from_slice(bytes, bincode::config::standard())
-            .map(|(s, _)| s)
-            .map_err(Error::from)
-    }
-}
-
-impl Sum for Scalar {
-    fn sum<I: Iterator<Item = Self>>(mut iter: I) -> Self {
-        let first = iter.next().expect("Iterator is empty");
-        match first {
-            Scalar::Secp256k1(first_scalar) => {
-                let sum = iter.fold(first_scalar, |acc, scalar| match scalar {
-                    Scalar::Secp256k1(s) => acc + s,
-                });
-                Scalar::Secp256k1(sum)
+        while let Some(scalar) = iter.next() {
+            if !sum.is_same_type(&scalar) {
+                return Err(Error::InconsistentVariants);
+            }
+            match (sum, scalar) {
+                (Scalar::Secp256k1(s1), Scalar::Secp256k1(s2)) => {
+                    sum = Scalar::Secp256k1(s1 + s2);
+                }
             }
         }
-    }
-}
-
-impl Sum for Point {
-    fn sum<I: Iterator<Item = Self>>(mut iter: I) -> Self {
-        let first = iter.next().expect("Iterator is empty");
-        match first {
-            Point::Secp256k1(first_point) => {
-                let sum = iter.fold(first_point, |acc, point| match point {
-                    Point::Secp256k1(p) => acc + p,
-                });
-                Point::Secp256k1(sum)
-            }
-        }
+        Ok(sum)
     }
 }
