+++ src/network/transport/libp2p_transport/protocols/gossipsub/payload.rs
@@ -7,7 +7,7 @@ use crate::{
     crypto::{
         dkg::Data,
         primitives::{
-            algebra::element::{Point, Scalar},
+            algebra::{Point, Scalar},
             vss::encrypted_share::EncryptedShares,
         },
     },
